move isDocumentLink to case utils
