+++ apps/consultation-portal/screens/Case/utils/index.ts
@@ -1,2 +1,3 @@
 export { default as isEmailValid } from './isEmailValid'
 export { default as renderDocFileName } from './renderDocFileName'
+export { default as isDocumentLink } from './isDocumentLink'
