+++ apps/consultation-portal/utils/helpers/index.ts
@@ -5,4 +5,3 @@ export * from './isObjectEmpty'
 export * from './createUUIDString'
 export * from './debounce'
 export * from './isAuthenticated'
-export { default as isDocumentLink } from './isDocumentLink'
